Branch test/e2e-order-swing-view:
Add the test methods testDeleteOrder, testDeleteOrderWhenOrderNoExists,
and testDeleteOrderWhenClientNoExists to verify GUI behavior when an
order is deleted correctly, or the client's order or order itself does
not exist. Add the test method testModifyOrder to verify GUI behavior
when an order is modified.

Progress on https://github.com/Tiezzi96/orders-management/issues/21