Branch test/e2e-order-swing-view:
Add the e2e test methods
testAddNewOrderForClientDifferentFromClientSelectedWithNewYear,
testAddNewOrderWithNoClientSelectedNoYearSelected,
testModifyOrderWhenNoYearsSelectedAndClientChangedIsDifferentFromClientSelected,
and testModifyOrderWhenNoYearsSelectedNoClientSelected to verify GUI
behavior when adding or updating an order with no year and/or no client
selected, or when the client for a modified order is different from the
currently selected client.

Add test method testShowAllOrdersClientsNoYearsSelected to verify that
the GUI shows all orders when no year is selected and the client is
deselected.

Progress on https://github.com/Tiezzi96/orders-management/issues/21